Move suspend functions back into Authgear.kt
